Seeding park data works (96 unique parks), remvoed import and switch to require in user models
