+++ package.json
@@ -7,7 +7,7 @@
     "start": "node server --ignore client",
     "develop": "concurrently \"cd server && npm run start\" \"cd client && npm start\"",
     "install": "cd server && npm i && cd ../client && npm i",
-    "seed": "cd server && npm run seed",
+    "seed": "cd server/config/ && node seed.js",
     "build": "cd client && npm run build"
   },
   "keywords": [],
