+++ server/config/connection.js
@@ -1,0 +1,11 @@
+const { connect, connection } = require('mongoose');
+
+const connectionString =
+  process.env.MONGODB_URI || 'mongodb://127.0.0.1:27017/NPS';
+
+connect(connectionString, {
+  useNewUrlParser: true,
+  useUnifiedTopology: true,
+});
+
+module.exports = connection;
