+++ server/models/index.js
@@ -1,4 +1,5 @@
 const User = require('./User');
 const Review = require('./Review')
+const Park = require('./Park')
 
-module.exports = { User, Review }
+module.exports = { User, Review, Park }
